Stashing process data revert functionality
Stashing changes as this would be best implemented along with the
new flow of processing data as "step 2.5", which will allow for
previews. With that flow in place, a user should not be allowed to
continue the first import or subsequent edits without processing,
both for the way the data is saved/can be reverted but also to
provide accurate previews.