Also filter view settings menus by those without defaults
All defaults will be saved if the dialog is saved. A possible issue
may be if the user makes a view visible but does not make any edits--
although that is a more global issue, to check if a view has all the
required field to display properly. No error checking was added here
as adding defaults for all views, visible or not, will bloat the
datasourcedescription.